More structure rework.
