+++ examples/draw_text/draw_text.ino
@@ -28,7 +28,7 @@
  */
 
 #include "ssd1306.h"
-#include "ssd1306_i2c_conf.h"
+#include "i2c/ssd1306_i2c_conf.h"
 
 /* Do not include wire.h for Attiny controllers */
 #ifdef SSD1306_WIRE_SUPPORTED
