+++ examples/menu_demo/menu_demo.ino
@@ -29,7 +29,7 @@
  */
 
 #include "ssd1306.h"
-#include "ssd1306_i2c_conf.h"
+#include "i2c/ssd1306_i2c_conf.h"
 #include "buttons.h"
 
 /* Do not include wire.h for Attiny controllers */
