+++ examples/sprite_pool/sprite_pool.ino
@@ -28,7 +28,7 @@
  */
 
 #include "ssd1306.h"
-#include "ssd1306_i2c_conf.h"
+#include "i2c/ssd1306_i2c_conf.h"
 #include "sprite_pool.h"
 
 /* Do not include wire.h for Attiny controllers */
