+++ examples/ssd1306_demo/ssd1306_demo.ino
@@ -28,8 +28,8 @@
  */
 
 #include "ssd1306.h"
-#include "ssd1306_i2c_conf.h"
-#include "ssd1306_spi_conf.h"
+#include "i2c/ssd1306_i2c_conf.h"
+#include "spi/ssd1306_spi_conf.h"
 #include "sova.h"
 #include "nano_gfx.h"
