+++ library.properties
@@ -1,5 +1,5 @@
 name=ssd1306
-version=1.2.6
+version=1.3.0
 author=Alexey Dynda
 maintainer=Alexey Dynda <alexey.dynda@gmail.com>
 sentence=SSD1306 i2c OLED Display driver, PCD8544 spi display driver
